requirements.txt
PyQt5
networkx
matplotlib
jinja2
python-whois
pymsfrpc
requests
